refactor: move connection.js to connection-manager.js
