+++ src/connection-manager.js
