+++ src/index.js
@@ -5,7 +5,7 @@ const EventEmitter = require('events').EventEmitter
 const each = require('async/each')
 const series = require('async/series')
 const TransportManager = require('./transport')
-const ConnectionManager = require('./connection')
+const ConnectionManager = require('./connection-manager')
 const getPeerInfo = require('./get-peer-info')
 const dial = require('./dial')
 const ProtocolMuxer = require('./protocol-muxer')
